requirement.txt
google-generativeai
Flask
Werkzeug
pandas
deepmultilingualpunctuation
yt-dlp
speechrecognition
